refactor: split http_interceptor into part-based proxy files
